Few tweaks in the rake file (to work in ruby 1.9) Updated to rake 0.9.2.2

+++ Rakefile
@@ -16,26 +16,27 @@ task :clobber do
   rm "Gemfile.lock" if File.exist?("Gemfile.lock")
 end
 
-namespace :rcov do
-  Cucumber::Rake::Task.new(:cucumber) do |t|    
-    t.rcov = true
-    t.rcov_opts = %w{--exclude osx\/objc,gems\/,spec\/,features\/ --aggregate coverage.data}
-    t.rcov_opts << %[-o "coverage"]
-  end
- 
-  RSpec::Core::RakeTask.new(:rspec) do |t|
-    t.rcov = true
-    t.rcov_opts = %w{--exclude osx\/objc,gems\/,spec\/,features\/}
-  end
- 
-  desc "Run both specs and features to generate aggregated coverage"
-  task :all do |t|
-    rm "coverage.data" if File.exist?("coverage.data")
-    Rake::Task["rcov:cucumber"].invoke
-    Rake::Task["rcov:rspec"].invoke
+if RUBY_VERSION.to_f <= 1.8
+  namespace :rcov do
+    Cucumber::Rake::Task.new(:cucumber) do |t|    
+      t.rcov = true
+      t.rcov_opts = %w{--exclude osx\/objc,gems\/,spec\/,features\/ --aggregate coverage.data}
+      t.rcov_opts << %[-o "coverage"]
+    end
+   
+    RSpec::Core::RakeTask.new(:rspec) do |t|
+      t.rcov = true
+      t.rcov_opts = %w{--exclude osx\/objc,gems\/,spec\/,features\/}
+    end
+   
+    desc "Run both specs and features to generate aggregated coverage"
+    task :all do |t|
+      rm "coverage.data" if File.exist?("coverage.data")
+      Rake::Task["rcov:cucumber"].invoke
+      Rake::Task["rcov:rspec"].invoke
+    end
   end
 end
-
 
 desc "Run rspec test"
 task :test do
